requirements.txt
Flask==2.2.5
feedparser==6.0.10
requests==2.31.0
APScheduler==3.10.1
python-dotenv==1.0.0